requirements.txt
# Hello Friends - Python Dependencies

# Core Framework
streamlit>=1.28.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# Data Processing
PyYAML>=6.0
pandas>=2.0.0

# File Handling
pathlib2>=2.3.7

# Future RAG Dependencies (commented out for now)
# openai>=1.0.0
# chromadb>=0.4.0
# langchain>=0.1.0
# sentence-transformers>=2.2.0
# pypdf>=3.0.0

# Development Dependencies
pytest>=7.0.0
black>=23.0.0
flake8>=6.0.0

# Optional: Enhanced UI components
streamlit-option-menu>=0.3.6
streamlit-extras>=0.3.0
